vid ex

+++ video_ex_strings/src/main.rs
@@ -3,5 +3,5 @@ fn main() {
     x.push_str(" Welcome to Rust!");
     // Must use collect otherwise mismatched types ^^^^^^^^^ expected `Vec<char>`, found `Chars<'_>`
     let y: Vec<char> = x.chars().collect();
-    println!("y1 = {}", y[1]);
+    println!("y1 = {}", y[1 as usize]); // as usize not neededs
 }
